removing networkx dependancy
